+++ NEWS.md
@@ -1,5 +1,10 @@
 ## v1.4 - Release Notes
 
+### Python
+
+* Python objects are now shown in the Environment Pane when `reticulate` REPL is active. (#6862)
+* Python objects can now be viewed using the Data Viewer and Object Explorer. (#6862)
+
 ### Plots
 
 * The default renderer used for the RStudio graphics device can now be customized. (#2142)
@@ -37,6 +42,7 @@
 
 * Fixed an issue where hovering mouse cursor over C++ completion popup would steal focus. (#5941)
 * Git integration now works properly for project names containing the '!' character. (#6160)
+* Fixed issue where loading the Rfast package could lead to session hangs. (#6645)
 * Fixed header resizing in Data Viewer (#1665)
 * Fixed resizing last column in Data Viewer (#2642)
 * Fixed inconsistencies in the resizing between a column and its header (#4361)